requirements.txt
arrow
selenium
discord.py